OpenGL 3.3 version of Ibex merged in for Windows replacing the older 2.x based version
- requires Assimp, GLFW to build now as well on VS2012, edit the ibex properties file to set the right paths

+++ IbexWin/utils.cpp
@@ -23,11 +23,11 @@ file_contents(const char *filename, GLint *length)
   }
 
   fseek(f, 0, SEEK_END);
-  *length = ftell(f);
+  *length = (GLint)ftell(f);
   fseek(f, 0, SEEK_SET);
 
   buffer = malloc(*length + 1);
-  *length = fread(buffer, 1, *length, f);
+  *length = (GLint)fread(buffer, 1, *length, f);
   fclose(f);
   ((char*) buffer)[*length] = '\0';
